requirements.txt
google-cloud-vision==3.10.0
paramiko==3.5.1
pillow==11.3.0
psycopg2-binary==2.9.10